# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doh Eain - Portal",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/units": {
            "get": {
                "tags": [
                    "Unit"
                ],
                "summary": "List",
                "description": "Get Unit List",
                "operationId": "unitList",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "per_page",
                                    "page",
                                    "sort_by"
                                ],
                                "properties": {
                                    "q": {
                                        "type": "string"
                                    },
                                    "per_page": {
                                        "type": "integer"
                                    },
                                    "page": {
                                        "type": "integer"
                                    },
                                    "sort_by": {
                                        "type": "string"
                                    },
                                    "sort_desc": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Unit"
                ],
                "summary": "Create",
                "description": "Create Unit",
                "operationId": "unitCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/units/{id}": {
            "get": {
                "tags": [
                    "Unit"
                ],
                "summary": "Get",
                "description": "Get Unit",
                "operationId": "unitGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Unit"
                ],
                "summary": "Update",
                "description": "Update Unit",
                "operationId": "unitUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Unit"
                ],
                "summary": "Delete",
                "description": "Delete Unit",
                "operationId": "unitDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}